pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wagtail-write-api"
version = "0.1.0"
description = "A read/write REST API for Wagtail CMS"
readme = "README.md"
license = "BSD-3-Clause"
requires-python = ">=3.10"
authors = [{ name = "Tom Dyson" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Wagtail",
    "License :: OSI Approved :: BSD License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "wagtail>=6.0",
    "django-ninja>=1.0",
    "djangorestframework",
    "markdown>=3.5",
    "markdownify>=0.11",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "factory-boy",
    "wagtail-factories",
    "ruff",
    "Pillow",
]

[project.urls]
Repository = "https://github.com/tomdyson/wagtail-write-api"

[tool.hatch.build.targets.wheel]
packages = ["src/wagtail_write_api"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "example_project.settings"
pythonpath = ["src", "example"]

[tool.ruff]
line-length = 99

[dependency-groups]
dev = [
    "zensical>=0.0.31",
]
